iterators: merge duplicate `iterate` methods for `Reverse`
The method bodies and state optional argument definition are exactly
the same.

The only difference between the two methods, is that the more specific
method, for `Reverse{<:AbstractArray}`, had a `@propagate_inbounds`
annotation. I simply deleted the more specific method, together with
the annotation. Deleting the annotation is consistent with the fact
that the `iterate(::AbstractArray)` method, in abstractarray.jl, has no
such annotation.